src/requirements.txt
FastAPI==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.26
python-dotenv==1.2.2
pydantic-settings==2.13.1
aiofiles==25.1.0
langchain==0.3.25
pyMupdf==1.27.2.2
langchain-community==0.3.24
